+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Modified helm recipe to work with any wooden trapdoor and wooden fence post
 * Modified seat recipe to work with any wooden stair and carpet
 * Modified pad recipe to work with any wooden fence post
+* Added a parrot perch
 
 ## 0.9.10-beta